Added preferences; Highscore is now saved

+++ core/src/main/kotlin/de/ash/xkay/ecs/systems/RemoveSystem.kt
@@ -36,7 +36,7 @@ class RemoveSystem(
             // Player died
             entity[PlayerComponent.mapper]?.let { player ->
                 eventManager.dispatchEvent(GameEvent.PlayerDeathEvent.apply {
-                    highscore = player.highscore.toLong()
+                    highscore = player.highscore.toInt()
                 })
             }
